demo1 tilemap init file, .tags ignore

+++ rex_proj/demo1_sim/inc/AnimatedGraphic.h
@@ -13,7 +13,7 @@ class AnimatedGraphic : public SDLGameObject
         virtual void clean();
 		virtual void load(const LoaderParams* pParams);
     private:
-        int m_animSpeed;
+        float m_animSpeed;
         int m_frameCount;
 };
